third_party/WebKit/LayoutTests/typedcssom/inlinestyle/inlineStylePropertyMap_delete.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<div id="testElement"></div>

<script>

// Delete
test(function() {
  testElement.style.width = '80px';
  assert_equals(testElement.styleMap.get('width').cssString, '80px');

  testElement.styleMap.delete('width');
  assert_equals(testElement.styleMap.get('width'), null);
  assert_equals(testElement.style.width, '');
}, "delete() removes the value from the property when set in element.style");

test(function() {
  testElement.styleMap.set('width', new SimpleLength(100, 'px'));
  assert_equals(testElement.styleMap.get('width').cssString, '100px');

  testElement.styleMap.delete('width');
  assert_equals(testElement.styleMap.get('width'), null);
  assert_equals(testElement.style.width, '');
}, "delete() removes the value from the property when set in element.styleMap");

test(function() {
  testElement.style.width = '90px';
  assert_equals(testElement.styleMap.get('width').cssString, '90px');

  testElement.styleMap.delete('WIdtH');
  assert_equals(testElement.styleMap.get('width'), null);
  assert_equals(testElement.style.width, '');
}, "delete() works when mixed case is used for the property name");

test(function() {
  assert_equals(testElement.styleMap.get('height'), null);
  testElement.styleMap.delete('height');
  assert_equals(testElement.styleMap.get('height'), null);
}, "delete() does nothing if the property isn't set");

test(function() {
  assert_throws(new TypeError(), function() {
    testElement.styleMap.delete('lemons');
  });
}, "Attempting to delete an invalid property throws");

// TODO(meade): Test deleting a property containing multiple values once that has been implemented.

</script>
